backend/requirements.txt
aiohttp==3.13.0
boto3==1.40.44
fastapi==0.116.1
psycopg[binary]==3.2.10
pydantic==2.11.7
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.43
tqdm==4.67.1
uvicorn==0.35.0
